refactor: moved design system components to a specific folder
