+++ react-client/src/components/design-system/index.ts
@@ -1,0 +1,5 @@
+export * from "./button/Button";
+export * from "./button/IconButton";
+export * from "./checkbox/Checkbox";
+export * from "./icon/Icon";
+export * from "./switch/Switch";
